Add files via upload
The book focuses on the synergy between nature and technology. It provides a detailed look at the formation and properties of basalt, followed by real-world applications ranging from Aerospace to automobile. With its clear explanations and real-life insights, the book serves as both an entry point for students and introduction guide for engineers and innovators.